initial user with authentication
