+++ api/app.py
@@ -1,8 +1,11 @@
 from fastapi import FastAPI
+from endpoints import user
 
 app = FastAPI()
+
+app.include_router(user.router)
 
 
 @app.get("/")
 async def root():
-    return {"message": "Hejloł"}
+    return {"message": "Hello Bigger Applications!"}
