+++ api/endpoints/user.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix="/user",
+    tags=["user"]
+)
+
+
