+++ api/models/__init__.py
@@ -1,20 +1,1 @@
-from pydantic import BaseModel
 
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    username: str | None = None
-
-
-class User(BaseModel):
-    username: str
-    email: str
-    disabled: bool
-
-
-class UserInDB(User):
-    hashed_password: str
